kernel/cpu: Rename cpu/x86/{apic.rs -> x2apic.rs}
The file only contains X2APIC functions, so rename it to reflect that.
